Refactor solid login widget
