+++ lib/src/solid/constants/ui.dart
@@ -238,135 +238,3 @@ class WebIdLayout {
 
   static const listPadding = EdgeInsets.fromLTRB(0, 5, 0, 5);
 }
-
-/// Button styles used in the Solid Login widget.
-
-// Colours for highlighted buttons.
-//
-// Do these need to be highligheted. By default the package should not highlight
-// them but if an app developer wants to then we should support that. (gjw
-// 20250422)
-//
-// The original alternatives were Color(0xFF00BCD4) and Colors.white for
-// register and Color(0xFF4CAF50) abd Colors.white for login. I find the colours
-// a bit distracting as a user. (gjw 20250422)
-
-const Color defaultButtonBackground = Colors.white;
-const Color defaultButtonForeground = Colors.black;
-
-const Color registerButtonBackground = defaultButtonBackground;
-const Color registerButtonForeground = defaultButtonForeground;
-const Color loginButtonBackground = defaultButtonBackground;
-const Color loginButtonForeground = defaultButtonForeground;
-
-const String defaultLoginButtonText = 'Login';
-const String defaultRegisterButtonText = 'Register';
-const String defaultInfoButtonText = 'Info';
-const String defaultContinueButtonText = 'Continue';
-const String defaultChangeKeyButtonText = 'Change Key';
-
-const String defaultServerTooltip = '''
-
-**Solid Server:** This text field contains the Solid server you will connect to
-where your data is hosted. It is also used as the base of the URI (Uniform
-Resource Identifier) that will be used for your WebID. A WebID is a
-decentralized identity that allows you to have a globally unique identifier for
-your data store.
-
-''';
-
-const String defaultLoginTooltip = '''
-
-**Login:** Tap here to log in to a Solid server to access you private data. You
-will be connected to the specified Solid server and you can then log in with
-your username and password. This app does not know your username/password. The
-app will will use a token from the server to establish your secure conenction.
-
-''';
-const String defaultRegisterTooltip = '''
-
-**Register:** Tap here to connect to your Solid server to register for an
-account. Once you have an account you will be able to save data onto your host
-server. You can connect to a Solid server of your choice, including your own, a
-free community supported server, a commercial server, or a government run
-server.
-
-''';
-
-const String defaultInfoTooltip = '''
-
-**Support:** Tap here to be taken to the app help and support documentation.
-
-''';
-
-const String defaultContinueTooltip = '''
-
-**Continue:** Tap here to continue on to the app without logging into your Solid
-server. The app will generally be able to save data locally or else prompt to
-log in to a Solid server when needed. No data will be shared beyond you local
-device until you connect to a SOlid server hosting your data.
-
-''';
-
-class ContinueButtonStyle {
-  const ContinueButtonStyle({
-    this.text = defaultContinueButtonText,
-    this.background = defaultButtonBackground,
-    this.foreground = defaultButtonForeground,
-    this.tooltip = defaultContinueTooltip,
-  });
-  final String text;
-  final Color background;
-  final Color foreground;
-  final String tooltip;
-}
-
-class ChangeKeyButtonStyle {
-  const ChangeKeyButtonStyle({
-    this.text = defaultChangeKeyButtonText,
-    this.background = defaultButtonBackground,
-    this.foreground = defaultButtonForeground,
-  });
-  final String text;
-  final Color background;
-  final Color foreground;
-}
-
-class LoginButtonStyle {
-  const LoginButtonStyle({
-    this.text = defaultLoginButtonText,
-    this.background = loginButtonBackground,
-    this.foreground = loginButtonForeground,
-    this.tooltip = defaultLoginTooltip,
-  });
-  final String text;
-  final Color background;
-  final Color foreground;
-  final String tooltip;
-}
-
-class RegisterButtonStyle {
-  const RegisterButtonStyle({
-    this.text = defaultRegisterButtonText,
-    this.background = registerButtonBackground,
-    this.foreground = registerButtonForeground,
-    this.tooltip = defaultRegisterTooltip,
-  });
-  final String text;
-  final Color background;
-  final Color foreground;
-  final String tooltip;
-}
-
-class InfoButtonStyle {
-  const InfoButtonStyle({
-    this.text = defaultInfoButtonText,
-    this.background = defaultButtonBackground,
-    this.foreground = defaultButtonForeground,
-    this.tooltip = defaultInfoTooltip,
-  });
-  final String text;
-  final Color background;
-  final Color foreground;
-  final String tooltip;
-}
